62-70 - Access control (Many chapters)
62 - Creating the Roles and Permissions Schema + UI
63 - Basic Access Control via Sessions
64 - Permissions Access Functions
65 - More Flexible Rule Based Functions
66 - Getting Meta - Roles based Roles and Hiding UI
67 - Cart and Order based Rules
68 - User and Field Based Permissions
69 - Product Image Permissions
70 - Creating a Gated Sign In Component

Something looked a bit funny in the instructions for chapter 62,
ended up pulling a bit more code than intended at once, so just
decided to follow the uncommited adjustments between chapters
62-70.

Made sure sell / orders / order pages are gated behind a "please
log in" component, and add conditional rendering of product
management UI. In the frontend we're missing a finer evaluation
of whether the user is capable of managing the given product(s)
before rendering the management UI buttons for those.

Ended up doing a temporary hack to ensure that I could create and
assign an admin role to the meant-to-be-admin user. The tutorial
doesn't seem to cover it, but in a never-released app probably the
right thing to do would have been to wipe the users and make sure
admin role is created along with first user.